test(ffi): extract smoke tests to src/tests.rs
Keeps lib.rs under the ~400-line module-size house rule (AGENTS.md):
the inline #[cfg(test)] module pushed it to 451; moving it to a sibling
file module brings lib.rs to 308. Stays a unit test (cdylib has no rlib
target for tests/ integration tests). Addresses PR review feedback.
